合并Dev/professional/sp3

See merge request chengdu/deepin-camera!1

+++ libcam/libcam_v4l2core/control_profile.c
@@ -216,19 +216,19 @@ int load_control_profile(v4l2_dev_t *vd, const char *filename)
 						   current->control.maximum == max &&
 						   current->control.step == step)
 						{
-							char str[max+1];
+                            char str[max+1];
 							char fmt[48];
 							sprintf(fmt,"ID{0x%%*x};CHK{%%*i:%%*i:%%*i:0}==STR{\"%%%is\"}", max);
-							sscanf(line, fmt, str);
+                            sscanf(line, fmt, str);
 
 							/*we are only scannig for max chars so this should never happen*/
-							if(strlen(str) > max) /*FIXME: should also check (minimum +N*step)*/
+                            if(strlen(str) > (size_t)max) /*FIXME: should also check (minimum +N*step)*/
 							{
                                 fprintf(stderr, "V4L2_CORE: (load_control_profile) string bigger than maximum buffer size (%i > %i)\n",
 									(int) strlen(str), max);
 								if(current->string)
 									free(current->string);
-								current->string = strndup(str, max); /*FIXME: does max includes '\0' ?*/
+                                current->string = strndup(str, (size_t)max); /*FIXME: does max includes '\0' ?*/
 							}
 							else
                             {
